add better match filter for rivens

+++ src-tauri/src/live_scraper/client.rs
@@ -2,7 +2,7 @@ use std::{
     collections::HashMap,
     sync::{
         atomic::{AtomicBool, Ordering},
-        Arc, Mutex,
+        Arc, Mutex, RwLock
     },
     time::Duration,
 };
@@ -29,6 +29,9 @@ use super::modules::{item::ItemModule, riven::RivenModule};
 #[derive(Clone)]
 pub struct LiveScraperClient {
     pub log_file: String,
+    component: String,
+    riven_module:Arc<RwLock<Option<RivenModule>>>,
+    item_module:Arc<RwLock<Option<ItemModule>>>,
     pub cache_querieds: Arc<Mutex<HashMap<String, DataFrame>>>,
     pub is_running: Arc<AtomicBool>,
     pub settings: Arc<Mutex<SettingsState>>,
@@ -52,6 +55,7 @@ impl LiveScraperClient {
     ) -> Self {
         LiveScraperClient {
             log_file: "live_scraper.log".to_string(),
+            component: "LiveScraper".to_string(),
             cache_querieds: Arc::new(Mutex::new(HashMap::new())),
             price_scraper,
             settings,
@@ -61,6 +65,8 @@ impl LiveScraperClient {
             db,
             cache,
             mh,
+            riven_module:Arc::new(RwLock::new(None)),
+            item_module:Arc::new(RwLock::new(None)),
         }
     }
     fn report_error(&self, error: AppError) {
@@ -73,7 +79,7 @@ impl LiveScraperClient {
             self.is_running.store(false, Ordering::SeqCst);
             crate::logger::dolog(
                 log_level.clone(),
-                component.as_str(),
+                format!("{}:{}", self.component, component).as_str(),
                 format!("{}, {}, {}", backtrace, cause, extra.to_string()).as_str(),
                 true,
                 Some(self.log_file.as_str()),
@@ -82,12 +88,34 @@ impl LiveScraperClient {
         } else {
             crate::logger::dolog(
                 log_level.clone(),
-                component.as_str(),
+                format!("{}:{}", self.component, component).as_str(),
                 format!("{}, {}, {}", backtrace, cause, extra.to_string()).as_str(),
                 true,
                 Some(self.log_file.as_str()),
             );
         }
+    }
+    pub fn debug(&self, id: &str, component: &str, msg: &str, file: Option<bool>) {
+        let settings = self.settings.lock().unwrap().clone();
+        if !settings.debug.contains(&"*".to_owned()) && !settings.debug.contains(&id.to_owned()) {
+            return;
+        }
+
+        if file.is_none() {
+            logger::debug(
+                format!("{}:{}", self.component, component).as_str(),
+                msg,
+                true,
+                None,
+            );
+            return;
+        }
+        logger::debug(
+            format!("{}:{}", self.component, component).as_str(),
+            msg,
+            true,
+            Some(&self.log_file),
+        );
     }
     pub fn stop_loop(&self) {
         self.is_running.store(false, Ordering::SeqCst);
@@ -105,23 +133,27 @@ impl LiveScraperClient {
         let db = self.db.lock()?.clone();
         // Reset riven stocks on start
         tauri::async_runtime::spawn(async move {
-            logger::info_con("LiveScraper", "Loop live scraper is started");
+            logger::info_con(&scraper.component, "Loop live scraper is started");
 
+            let mut settings = scraper.settings.lock().unwrap().clone();
             scraper.send_message("riven.reset", None);
             db.stock_riven().reset_listed_price().await.unwrap();
             scraper.send_message("item.reset", None);
-            db.stock_item().reset_listed_price().await.unwrap();
-            scraper
-                .item()
-                .delete_all_orders(OrderMode::Both)
-                .await
-                .unwrap();
+            if settings.live_scraper.stock_item.auto_delete {
+                db.stock_item().reset_listed_price().await.unwrap();
+                scraper
+                    .item()
+                    .delete_all_orders(OrderMode::Both)
+                    .await
+                    .unwrap();
+            }
 
-            let riven_interval = 60;
+            let riven_interval = 5;
             let mut current_riven_interval = riven_interval.clone();
 
             while is_running.load(Ordering::SeqCst) && forced_stop.load(Ordering::SeqCst) {
-                let settings = scraper.settings.lock().unwrap().clone();
+                settings = scraper.settings.lock().unwrap().clone();
+
                 if (settings.live_scraper.stock_mode == StockMode::Riven
                     || settings.live_scraper.stock_mode == StockMode::All)
                     && current_riven_interval >= riven_interval
@@ -133,6 +165,7 @@ impl LiveScraperClient {
                         Err(e) => scraper.report_error(e),
                     }
                 }
+
                 if settings.live_scraper.stock_mode == StockMode::Item
                     || settings.live_scraper.stock_mode == StockMode::All
                 {
@@ -146,15 +179,41 @@ impl LiveScraperClient {
                 tokio::time::sleep(Duration::from_secs(1)).await;
             }
             scraper.send_message("", None);
-            logger::info_con("LiveScraper", "Loop live scraper is stopped");
+            logger::info_con(&scraper.component, "Loop live scraper is stopped");
         });
         Ok(())
     }
     pub fn item(&self) -> ItemModule {
-        ItemModule { client: self }
+        // Lazily initialize ItemModule if not already initialized
+        if self.item_module.read().unwrap().is_none() {
+            *self.item_module.write().unwrap() = Some(ItemModule::new(self.clone()).clone());
+        }
+
+        // Unwrapping is safe here because we ensured the item_module is initialized
+        self.item_module.read().unwrap().as_ref().unwrap().clone()
     }
+    pub fn update_item_module(&self, module: ItemModule) {
+        // Update the stored ItemModule
+        *self.item_module.write().unwrap() = Some(module);
+    }
+    // pub fn item(&self) -> ItemModule {
+    //     ItemModule {
+    //         client: self,
+    //         debug_id: "live_scraper_item".to_string(),
+    //     }
+    // }
     pub fn riven(&self) -> RivenModule {
-        RivenModule { client: self }
+        // Lazily initialize ItemModule if not already initialized
+        if self.riven_module.read().unwrap().is_none() {
+            *self.riven_module.write().unwrap() = Some(RivenModule::new(self.clone()).clone());
+        }
+
+        // Unwrapping is safe here because we ensured the riven_module is initialized
+        self.riven_module.read().unwrap().as_ref().unwrap().clone()
+    }
+    pub fn update_riven_module(&self, module: RivenModule) {
+        // Update the stored ItemModule
+        *self.riven_module.write().unwrap() = Some(module);
     }
 
     pub fn send_message(&self, i18n_key: &str, data: Option<serde_json::Value>) {
